@@ -1,18 +1,18 @@
 [package]
 name = "taxonomy"
-version = "0.7.1"
-authors = ["Vincent Prouillet & One Codex <vincent@onecodex.com>"]
+version = "0.8.0"
+authors = ["Vincent Prouillet <vincent.prouillet@invitae.com>"]
 description = "Routines for loading, saving, and manipulating taxonomic trees"
 keywords = ["taxonomy", "bioinformatics"]
 categories = ["command-line-utilities", "science"]
 repository = "https://github.com/onecodex/taxonomy"
 license = "MIT"
 readme = "./README.md"
-edition = "2018"
+edition = "2021"
 
 [dependencies]
 memchr = "2.2.1"
-pyo3 = { version = "0.14", optional = true }
+pyo3 = { version = "0.15", optional = true }
 quick-xml = "0.22"
 serde = { version = "1.0.104", features = ["derive"] }
 serde_json = "1.0.44"
@@ -28,6 +28,7 @@ bench = false
 [dev-dependencies]
 criterion = "0.3"
 rand = "0.8"
+tempfile = "3.3"
 
 [[bench]]
 name = "taxonomy"
