@@ -1,33 +1,73 @@
-//! Implementation of the Python API.
-//!
-//! Only enabled when `cargo build --features python` is run or when the
-//! Python API is build with `python setup.py develop`.
 use std::collections::HashMap;
-use std::fs::File;
 use std::io::Cursor;
-use std::str::FromStr;
 
-use pyo3::class::{PyIterProtocol, PyMappingProtocol, PyObjectProtocol, PySequenceProtocol};
-use pyo3::exceptions::PyKeyError;
 use pyo3::prelude::*;
-use pyo3::types::{IntoPyDict, PyBytes, PyDict, PyType};
-use pyo3::{create_exception, wrap_pyfunction, wrap_pymodule};
+use pyo3::types::{PyBytes, PyDict, PyList, PyType};
+use pyo3::{
+    create_exception, PyIterProtocol, PyMappingProtocol, PyObjectProtocol, PySequenceProtocol,
+};
+use serde_json::Value;
 
-use crate::base::{GeneralTaxonomy, IntTaxId};
-use crate::edit::{prune_away, prune_to};
-use crate::formats::json::{load_json, save_json};
-use crate::formats::ncbi::load_ncbi;
-use crate::formats::newick::{load_newick, save_newick};
-use crate::formats::phyloxml::load_phyloxml;
+use crate::json::JsonFormat;
 use crate::rank::TaxRank;
-use crate::taxonomy::Taxonomy as TaxTrait;
-use crate::weights::{maximum_weighted_path as tax_mwp, rollup_weights as tax_rw};
+use crate::Taxonomy as TaxonomyTrait;
+use crate::{json, ncbi, newick, phyloxml, GeneralTaxonomy};
+use pyo3::basic::CompareOp;
+use pyo3::exceptions::PyKeyError;
+use std::ops::Deref;
+use std::str::FromStr;
 
 create_exception!(taxonomy, TaxonomyError, pyo3::exceptions::PyException);
 
+// Avoid some boilerplate with the error handling
+macro_rules! py_try {
+    ($call:expr) => {
+        $call.map_err(|e| PyErr::new::<TaxonomyError, _>(format!("{}", e)))?
+    };
+    ($call:expr, $msg:expr) => {
+        $call.map_err(|_| PyErr::new::<TaxonomyError, _>($msg.to_owned()))?
+    };
+}
+
+// child_readcount = sum([graph.nodes[int(tid)]["readcount"] for tid in child_tax_ids])
+
+fn json_value_to_pyobject(val: &Value) -> PyObject {
+    let gil_guard = Python::acquire_gil();
+    let py = gil_guard.python();
+    match val {
+        Value::Null => py.None(),
+        Value::Bool(b) => b.to_object(py),
+        Value::Number(n) => {
+            if let Some(n1) = n.as_i64() {
+                return n1.to_object(py);
+            }
+            n.as_f64().unwrap().to_object(py)
+        }
+        Value::String(s) => s.to_object(py),
+        Value::Array(arr) => {
+            let pylist = PyList::empty(py);
+            for v in arr {
+                pylist
+                    .append(json_value_to_pyobject(v))
+                    .expect("can add items to list");
+            }
+            pylist.to_object(py)
+        }
+        Value::Object(obj) => {
+            let pydict = PyDict::new(py);
+            for (key, val) in obj.iter() {
+                pydict
+                    .set_item(key, json_value_to_pyobject(val))
+                    .expect("can add items to dict");
+            }
+            pydict.to_object(py)
+        }
+    }
+}
+
 /// The data returned when looking up a taxonomy by id or by name
 #[pyclass]
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Eq, PartialEq)]
 pub struct TaxonomyNode {
     #[pyo3(get)]
     id: String,
@@ -37,92 +77,95 @@ pub struct TaxonomyNode {
     parent: Option<String>,
     #[pyo3(get)]
     rank: String,
+    // Ideally this would be private
+    extra: HashMap<String, Value>,
 }
 
 #[pyproto]
 impl PyObjectProtocol for TaxonomyNode {
-    fn __repr__(&self) -> PyResult<String> {
-        Ok(format!(
-            "<TaxonomyNode (id={} rank=\"{}\" name=\"{}\"))>",
-            self.id, self.rank, self.name
-        ))
+    fn __richcmp__(&self, other: PyRef<TaxonomyNode>, op: CompareOp) -> Py<PyAny> {
+        let py = other.py();
+        match op {
+            CompareOp::Eq => (self == other.deref()).into_py(py),
+            CompareOp::Ne => (self != other.deref()).into_py(py),
+            _ => py.NotImplemented(),
+        }
+    }
+}
+
+#[pyproto]
+impl PyMappingProtocol for TaxonomyNode {
+    fn __getitem__(&self, obj: &PyAny) -> PyResult<PyObject> {
+        let key: &str = obj.extract()?;
+        let gil_guard = Python::acquire_gil();
+        let py = gil_guard.python();
+        match key {
+            "id" => Ok(self.id.to_object(py)),
+            "name" => Ok(self.name.to_object(py)),
+            "parent" => Ok(self.parent.to_object(py)),
+            "rank" => Ok(self.rank.to_object(py)),
+            _ => {
+                if self.extra.contains_key(key) {
+                    Ok(json_value_to_pyobject(self.extra.get(key).unwrap()))
+                } else {
+                    return Err(PyKeyError::new_err(format!("Key {} not found", key)));
+                }
+            }
+        }
     }
 }
 
 /// The Taxonomy object provides the primary interface for exploring a
-/// biological taxonomy. Iterating over the Taxonomy returns all the taxonomy
-/// ids in pre-order fashion and indexing into the Taxonomy object with a
-/// taxonomy id will return a tuple of the name and rank of that id.
+/// biological taxonomy.
 #[pyclass]
 #[derive(Debug)]
 pub struct Taxonomy {
-    t: GeneralTaxonomy,
-}
-
-// Avoid some boilerplate with the error handling
-macro_rules! py_try {
-    ($call:expr) => {
-        $call.map_err(|e| PyErr::new::<TaxonomyError, _>(format!("{}", e)))?
-    };
+    tax: GeneralTaxonomy,
 }
 
 /// Some private fns we don't want to share in Python but that make the Python code easier to
 /// write
 impl Taxonomy {
-    pub(crate) fn get_int_id(&self, key: &str) -> PyResult<usize> {
-        self.t
-            .to_internal_id(key)
-            .map_err(|_| PyErr::new::<PyKeyError, _>("Tax ID is not in taxonomy"))
-    }
-
-    pub(crate) fn get_name(&self, key: &str) -> PyResult<&str> {
-        let name = py_try!(self.t.name(key));
+    pub(crate) fn get_name(&self, tax_id: &str) -> PyResult<&str> {
+        let name = py_try!(self.tax.name(tax_id));
         Ok(name)
     }
 
-    pub(crate) fn get_rank(&self, key: &str) -> PyResult<&str> {
-        let rank = py_try!(self.t.rank(key)).to_ncbi_rank();
+    pub(crate) fn get_ncbi_rank(&self, tax_id: &str) -> PyResult<&str> {
+        let rank = py_try!(self.tax.rank(tax_id)).to_ncbi_rank();
         Ok(rank)
     }
 
-    pub(crate) fn as_node(&self, key: &str) -> PyResult<TaxonomyNode> {
-        let name = self.get_name(key)?;
-        let rank = self.get_rank(key)?;
-        let parent = py_try!(self.t.parent(key)).map(|(p, _)| p.to_string());
+    pub(crate) fn as_node(&self, tax_id: &str) -> PyResult<TaxonomyNode> {
+        let name = self.get_name(tax_id)?;
+        let rank = self.get_ncbi_rank(tax_id)?;
+        let parent = py_try!(self.tax.parent(tax_id)).map(|(p, _)| p.to_string());
+        let extra = py_try!(self.tax.data(tax_id));
 
         Ok(TaxonomyNode {
-            id: key.to_string(),
+            id: tax_id.to_string(),
             name: name.to_string(),
             rank: rank.to_string(),
+            extra: extra.clone(),
             parent,
         })
     }
 }
 
-#[pyproto]
-impl PyObjectProtocol for Taxonomy {
-    fn __repr__(&self) -> PyResult<String> {
-        Ok(format!(
-            "<Taxonomy ({} nodes)>",
-            TaxTrait::<IntTaxId, _>::len(&self.t)
-        ))
-    }
-}
-
 #[pymethods]
 impl Taxonomy {
-    /// from_json(cls, value: str, /, path: List[str])
+    /// from_json(cls, value: str, /, json_pointer: str)
     /// --
     ///
     /// Load a Taxonomy from a JSON-encoded string. The format can either be
     /// of the tree or node_link_data types and will be automatically detected.
     /// If `path` is specified, the JSON will be traversed to that sub-object
-    /// before being parsed as a taxonomy.
+    /// before being parsed as a taxonomy. `path` has to be a valid JSON path string.
     #[classmethod]
-    fn from_json(_cls: &PyType, value: &str, path: Option<Vec<&str>>) -> PyResult<Taxonomy> {
+    fn from_json(_cls: &PyType, value: &str, json_pointer: Option<&str>) -> PyResult<Taxonomy> {
         let mut c = Cursor::new(value);
-        let t = py_try!(load_json(&mut c, path.as_deref()));
-        Ok(Taxonomy { t })
+        let tax = py_try!(json::load(&mut c, json_pointer));
+        Ok(Taxonomy { tax })
     }
 
     /// from_newick(cls, value: str)
@@ -132,22 +175,19 @@ impl Taxonomy {
     #[classmethod]
     fn from_newick(_cls: &PyType, value: &str) -> PyResult<Taxonomy> {
         let mut c = Cursor::new(value);
-        let t = py_try!(load_newick(&mut c));
-        Ok(Taxonomy { t })
+        let tax = py_try!(newick::load(&mut c));
+        Ok(Taxonomy { tax })
     }
 
-    /// from_ncbi(cls, nodes_path: str, names_path: str)
+    /// from_ncbi(cls, dump_dir: str)
     /// --
     ///
-    /// Load a Taxonomy from a pair of NCBI dump files. The paths specified are
-    /// to the individual files in the NCBI taxonomy directory (e.g. nodes.dmp
-    /// and names.dmp).
+    /// Load a Taxonomy from a directory.
+    /// The directory must contain the `nodes.dmp` and `names.dmp` files.
     #[classmethod]
-    fn from_ncbi(_cls: &PyType, nodes_path: &str, names_path: &str) -> PyResult<Taxonomy> {
-        let nodes_file = File::open(nodes_path)?;
-        let names_file = File::open(names_path)?;
-        let t = py_try!(load_ncbi(nodes_file, names_file));
-        Ok(Taxonomy { t })
+    fn from_ncbi(_cls: &PyType, dump_dir: &str) -> PyResult<Taxonomy> {
+        let tax = py_try!(ncbi::load(dump_dir));
+        Ok(Taxonomy { tax })
     }
 
     /// from_phyloxml(cls, value: str)
@@ -159,28 +199,37 @@ impl Taxonomy {
     #[classmethod]
     fn from_phyloxml(_cls: &PyType, value: &str) -> PyResult<Taxonomy> {
         let mut c = Cursor::new(value);
-        let t = py_try!(load_phyloxml(&mut c));
-        Ok(Taxonomy { t })
+        let tax = py_try!(phyloxml::load(&mut c));
+        Ok(Taxonomy { tax })
     }
 
-    /// to_json(self, /, as_node_link_data: bool)
+    /// to_json_tree(self)
     /// --
     ///
-    /// Export a Taxonomy as a JSON-encoded byte string. By default, the JSON format
-    /// is a tree format unless the `as_node_link_data` parameter is set to True.
-    #[args(as_node_link_data = false)]
-    fn to_json(&self, as_node_link_data: bool) -> PyResult<PyObject> {
-        let mut s = Vec::new();
-        py_try!(save_json::<&str, _, _, _>(
-            &self.t,
-            &mut s,
-            None,
-            as_node_link_data
-        ));
-
+    /// Export a Taxonomy as a JSON-encoded byte string in a tree format
+    fn to_json_tree(&self) -> PyResult<PyObject> {
+        let mut bytes = Vec::new();
+        py_try!(json::save(&mut bytes, &self.tax, JsonFormat::Tree, None));
         let gil = Python::acquire_gil();
         let py = gil.python();
-        Ok(PyBytes::new(py, &s).into())
+        Ok(PyBytes::new(py, &bytes).into())
+    }
+
+    /// to_json_node_links(self)
+    /// --
+    ///
+    /// Export a Taxonomy as a JSON-encoded byte string in a node link format
+    fn to_json_node_links(&self) -> PyResult<PyObject> {
+        let mut bytes = Vec::new();
+        py_try!(json::save(
+            &mut bytes,
+            &self.tax,
+            JsonFormat::NodeLink,
+            None
+        ));
+        let gil = Python::acquire_gil();
+        let py = gil.python();
+        Ok(PyBytes::new(py, &bytes).into())
     }
 
     /// to_newick(self)
@@ -188,12 +237,11 @@ impl Taxonomy {
     ///
     /// Export a Taxonomy as a Newick-encoded byte string.
     fn to_newick(&self) -> PyResult<PyObject> {
-        let mut s = Vec::new();
-        py_try!(save_newick::<&str, _, _, _>(&self.t, &mut s, None));
-
+        let mut bytes = Vec::new();
+        py_try!(newick::save(&mut bytes, &self.tax, Some(self.tax.root())));
         let gil = Python::acquire_gil();
         let py = gil.python();
-        Ok(PyBytes::new(py, &s).into())
+        Ok(PyBytes::new(py, &bytes).into())
     }
 
     /// node(self, tax_id: str) -> Optional[TaxonomyNode]
@@ -209,7 +257,7 @@ impl Taxonomy {
     ///
     /// Find a node by its name, Raises an exception if not found.
     fn find_by_name(&self, name: &str) -> Option<TaxonomyNode> {
-        if let Ok(tax_id) = self.t.find_by_name(name) {
+        if let Some(tax_id) = self.tax.find_by_name(name) {
             self.as_node(tax_id).ok()
         } else {
             None
@@ -230,7 +278,7 @@ impl Taxonomy {
     ) -> PyResult<(Option<TaxonomyNode>, Option<f32>)> {
         let parent_res = if let Some(rank) = at_rank {
             if let Ok(rank) = TaxRank::from_str(rank) {
-                self.t.parent_at_rank(tax_id, rank)
+                self.tax.parent_at_rank(tax_id, rank)
             } else {
                 return Err(PyErr::new::<TaxonomyError, _>(format!(
                     "Rank {} could not be understood",
@@ -238,7 +286,7 @@ impl Taxonomy {
                 )));
             }
         } else {
-            self.t.parent(tax_id)
+            self.tax.parent(tax_id)
         };
 
         if let Ok(Some((id, distance))) = parent_res {
@@ -247,6 +295,7 @@ impl Taxonomy {
             Ok((None, None))
         }
     }
+
     /// parent(self, tax_id: str, /, at_rank: str)
     /// --
     ///
@@ -264,7 +313,7 @@ impl Taxonomy {
     ///
     /// Return a list of child taxonomy nodes from the node id provided.
     fn children(&self, tax_id: &str) -> PyResult<Vec<TaxonomyNode>> {
-        let children: Vec<&str> = py_try!(self.t.children(tax_id));
+        let children: Vec<&str> = py_try!(self.tax.children(tax_id));
         let mut res = Vec::with_capacity(children.len());
         for key in children {
             let child = self.node(key);
@@ -287,7 +336,7 @@ impl Taxonomy {
     /// Return a list of all the parent taxonomy nodes of the node id provided
     /// (including that node itself).
     fn lineage(&self, tax_id: &str) -> PyResult<Vec<TaxonomyNode>> {
-        let lineage: Vec<&str> = py_try!(self.t.lineage(tax_id));
+        let lineage: Vec<&str> = py_try!(self.tax.lineage(tax_id));
         let mut res = Vec::with_capacity(lineage.len());
         for key in lineage {
             let ancestor = self.node(key);
@@ -308,6 +357,7 @@ impl Taxonomy {
     /// --
     ///
     /// Return a list of all the parent taxonomy nodes of the node id provided.
+    /// It is equivalent to `lineage` except it doesn't include itself
     fn parents(&self, tax_id: &str) -> PyResult<Vec<TaxonomyNode>> {
         let mut lineage = self.lineage(tax_id)?;
         lineage.drain(..1);
@@ -319,7 +369,7 @@ impl Taxonomy {
     ///
     /// Return the lowest common ancestor of two taxonomy nodes.
     fn lca(&self, id1: &str, id2: &str) -> PyResult<Option<TaxonomyNode>> {
-        let lca_id = py_try!(self.t.lca(id1, id2));
+        let lca_id = py_try!(self.tax.lca(id1, id2));
         Ok(self.node(lca_id))
     }
 
@@ -330,14 +380,14 @@ impl Taxonomy {
     ///  - only the nodes in `keep` and their parents if provided
     ///  - all of the nodes except those in remove and their children if provided
     fn prune(&self, keep: Option<Vec<&str>>, remove: Option<Vec<&str>>) -> PyResult<Taxonomy> {
-        let mut tax = self.t.clone();
+        let mut tax = self.tax.clone();
         if let Some(k) = keep {
-            tax = py_try!(prune_to(&tax, &k, false));
+            tax = py_try!(tax.prune_to(&k, false));
         }
         if let Some(r) = remove {
-            tax = py_try!(prune_away(&tax, &r));
+            tax = py_try!(tax.prune_away(&r));
         }
-        Ok(Taxonomy { t: tax })
+        Ok(Taxonomy { tax })
     }
 
     /// remove_node(self, tax_id: str)
@@ -345,8 +395,7 @@ impl Taxonomy {
     ///
     /// Remove the node from the tree.
     fn remove_node(&mut self, tax_id: &str) -> PyResult<()> {
-        let int_id = self.get_int_id(tax_id)?;
-        py_try!(self.t.remove(int_id));
+        py_try!(self.tax.remove(tax_id));
         Ok(())
     }
 
@@ -355,8 +404,7 @@ impl Taxonomy {
     ///
     /// Add a new node to the tree at the parent provided.
     fn add_node(&mut self, parent_id: &str, tax_id: &str) -> PyResult<()> {
-        let int_id = self.get_int_id(parent_id)?;
-        py_try!(self.t.add(int_id, tax_id));
+        py_try!(self.tax.add(parent_id, tax_id));
         Ok(())
     }
 
@@ -372,95 +420,98 @@ impl Taxonomy {
         parent_id: Option<&str>,
         parent_distance: Option<f32>,
     ) -> PyResult<()> {
-        let int_id = self.get_int_id(tax_id)?;
+        let idx = py_try!(self.tax.to_internal_index(tax_id));
 
         if let Some(r) = rank {
-            self.t.ranks[int_id] = TaxRank::from_str(r)
-                .map_err(|_| PyErr::new::<TaxonomyError, _>("Rank could not be understood"))?;
+            self.tax.ranks[idx] = py_try!(TaxRank::from_str(r), "Rank could not be understood");
         }
         if let Some(n) = name {
-            self.t.names[int_id] = n.to_string();
+            self.tax.names[idx] = n.to_string();
         }
         if let Some(p) = parent_id {
-            if int_id == TaxTrait::<IntTaxId, _>::root(&self.t) {
-                return Err(PyErr::new::<TaxonomyError, _>("Root has no parent"));
+            if tax_id == self.tax.root() {
+                return Err(PyErr::new::<TaxonomyError, _>("Root cannot have a parent"));
             }
-            let parent = self
-                .t
-                .to_internal_id(p)
-                .map_err(|_| PyErr::new::<PyKeyError, _>("New parent ID is not in taxonomy"))?;
-            if TaxTrait::<IntTaxId, _>::lineage(&self.t, parent)
-                .map_err(|_| PyErr::new::<TaxonomyError, _>("New parent has bad lineage?"))?
-                .contains(&int_id)
-            {
+            let parent = py_try!(self.tax.parent(p))
+                .ok_or_else(|| PyErr::new::<PyKeyError, _>("New parent ID is not in taxonomy"))?
+                .0;
+            let lineage = py_try!(self.tax.lineage(parent), "New parent has bad lineage?");
+            if lineage.contains(&tax_id) {
                 return Err(PyErr::new::<TaxonomyError, _>(
                     "Node can not be moved to its child",
                 ));
             }
-            self.t.parent_ids[int_id] = parent;
+            self.tax.parent_ids[idx] = py_try!(self.tax.to_internal_index(parent));
         }
+
         if let Some(p) = parent_distance {
-            if int_id == TaxTrait::<IntTaxId, _>::root(&self.t) {
-                return Err(PyErr::new::<TaxonomyError, _>("Root has no parent"));
+            if tax_id == self.tax.root() {
+                return Err(PyErr::new::<TaxonomyError, _>("Root cannot have a parent"));
             }
-            self.t.parent_dists[int_id] = p;
+            self.tax.parent_distances[idx] = p;
         }
+
         Ok(())
     }
 
     #[getter]
     fn root(&self) -> TaxonomyNode {
-        let key: &str = self.t.root();
+        let key: &str = self.tax.root();
         self.as_node(key).unwrap()
+    }
+}
+
+#[pyproto]
+impl PyObjectProtocol for Taxonomy {
+    fn __repr__(&self) -> PyResult<String> {
+        Ok(format!("<Taxonomy ({} nodes)>", self.tax.len()))
     }
 }
 
 #[pyproto]
 impl PyMappingProtocol for Taxonomy {
     fn __len__(&self) -> PyResult<usize> {
-        Ok(TaxTrait::<IntTaxId, _>::len(&self.t))
+        Ok(self.tax.len())
     }
 
     // TODO: way to set name and rank
     // TODO: also way to set parent and parent distance?
 
-    fn __getitem__(&self, key: &str) -> PyResult<TaxonomyNode> {
-        self.as_node(key)
+    fn __getitem__(&self, tax_id: &str) -> PyResult<TaxonomyNode> {
+        self.as_node(tax_id)
     }
 
-    fn __delitem__(&mut self, key: &str) -> PyResult<()> {
-        let int_id = self.get_int_id(key)?;
-        py_try!(self.t.remove(int_id));
-        Ok(())
+    fn __delitem__(&mut self, tax_id: &str) -> PyResult<()> {
+        Ok(py_try!(self.tax.remove(tax_id)))
     }
 }
 
 #[pyproto]
 impl PySequenceProtocol for Taxonomy {
-    fn __contains__(&self, key: &str) -> PyResult<bool> {
-        Ok(self.get_int_id(key).is_ok())
+    fn __contains__(&self, tax_id: &str) -> PyResult<bool> {
+        Ok(self.tax.to_internal_index(tax_id).is_ok())
     }
 }
 
 #[pyproto]
 impl PyIterProtocol for Taxonomy {
     fn __iter__(slf: PyRefMut<Self>) -> PyResult<TaxonomyIterator> {
-        let root = TaxTrait::<IntTaxId, _>::root(&slf.t);
+        let root = slf.tax.root();
+        let root_idx = slf.tax.to_internal_index(root).unwrap();
         let gil_guard = Python::acquire_gil();
         let py = gil_guard.python();
         Ok(TaxonomyIterator {
             t: slf.into_py(py),
-            nodes_left: vec![root],
+            nodes_left: vec![root_idx],
             visited_nodes: Vec::new(),
         })
     }
 }
-
 #[pyclass]
 pub struct TaxonomyIterator {
     t: PyObject,
-    visited_nodes: Vec<IntTaxId>,
-    nodes_left: Vec<IntTaxId>,
+    visited_nodes: Vec<usize>,
+    nodes_left: Vec<usize>,
 }
 
 #[pyproto]
@@ -485,10 +536,15 @@ impl PyIterProtocol for TaxonomyIterator {
             } else {
                 slf.visited_nodes.push(cur_node.clone());
                 let tax: PyRef<Taxonomy> = slf.t.extract(py)?;
+                let cur_node_str = tax.tax.from_internal_index(cur_node).unwrap();
                 let children = tax
-                    .t
-                    .children(cur_node)
+                    .tax
+                    .children(cur_node_str)
                     .map_err(|e| PyErr::new::<TaxonomyError, _>(format!("{}", e)))?;
+                let children: Vec<_> = children
+                    .into_iter()
+                    .map(|c| tax.tax.to_internal_index(c).unwrap())
+                    .collect();
                 // drop the reference to tax, we don't need it anymore
                 drop(tax);
                 if !children.is_empty() {
@@ -499,8 +555,8 @@ impl PyIterProtocol for TaxonomyIterator {
             if node_visited == !traverse_preorder {
                 let tax: PyRef<Taxonomy> = slf.t.extract(py)?;
                 return Ok(Some(
-                    tax.t
-                        .from_internal_id(node)
+                    tax.tax
+                        .from_internal_index(node)
                         .map_err(|e| PyErr::new::<TaxonomyError, _>(format!("{}", e)))?
                         .to_string(),
                 ));
@@ -509,463 +565,11 @@ impl PyIterProtocol for TaxonomyIterator {
     }
 }
 
-/// maximum_weighted_path(tax: Taxonomy, weights: Dict[str, float])
-/// --
-///
-#[pyfunction]
-fn maximum_weighted_path(tax: &Taxonomy, weights: &PyDict) -> PyResult<Option<(String, f32)>> {
-    // TODO: remove this once https://github.com/PyO3/pyo3/pull/702 lands in a release
-    let mut hash_weights: HashMap<&str, f32> = HashMap::new();
-    for (k, v) in weights.iter() {
-        hash_weights.insert(k.extract()?, v.extract()?);
-    }
-    let weights = hash_weights;
-
-    let mwp = tax_mwp(&tax.t, &weights, false)
-        .map_err(|e| PyErr::new::<TaxonomyError, _>(format!("{}", e)))?;
-
-    Ok(mwp.map(|x| (x.0.to_string(), x.1)))
-}
-
-/// rollup_weights(tax: Taxonomy, weights: Dict[str, float])
-/// --
-///
-#[pyfunction]
-fn rollup_weights(tax: &Taxonomy, weights: &PyDict) -> PyResult<PyObject> {
-    // TODO: remove this once https://github.com/PyO3/pyo3/pull/702 lands in a release
-    let mut hash_weights: HashMap<&str, f32> = HashMap::new();
-    for (k, v) in weights.iter() {
-        hash_weights.insert(k.extract()?, v.extract()?);
-    }
-    let weights = hash_weights;
-
-    let gil = Python::acquire_gil();
-    let py = gil.python();
-
-    let rolled: Vec<(&str, f32)> =
-        tax_rw(&tax.t, &weights).map_err(|e| PyErr::new::<TaxonomyError, _>(format!("{}", e)))?;
-    Ok(rolled.into_py_dict(py).into())
-}
-
-/// Functions related to calculations using node weights
-#[pymodule]
-fn weights(_py: Python, m: &PyModule) -> PyResult<()> {
-    m.add_wrapped(wrap_pyfunction!(maximum_weighted_path))?;
-    m.add_wrapped(wrap_pyfunction!(rollup_weights))?;
-
-    Ok(())
-}
-
 /// The taxonomy module
 #[pymodule]
 fn taxonomy(py: Python, m: &PyModule) -> PyResult<()> {
     m.add_class::<Taxonomy>()?;
     m.add("TaxonomyError", py.get_type::<TaxonomyError>())?;
 
-    let m_weights = wrap_pymodule!(weights)(py);
-    m.add("weights", &m_weights)?;
-    // this allows `from taxonomy.weights import ...`
-    // see https://github.com/PyO3/pyo3/issues/471
-    py.import("sys")?
-        .dict()
-        .get_item("modules")
-        .unwrap()
-        .set_item("taxonomy.weights", m_weights)?;
-
     Ok(())
-}
-
-#[cfg(test)]
-mod test {
-    use pyo3::prelude::*;
-    use pyo3::types::PyDict;
-
-    use crate::base::test::create_example;
-    use crate::taxonomy::Taxonomy as TaxTrait;
-
-    use super::{weights, Taxonomy, TaxonomyNode};
-    use crate::TaxRank;
-
-    fn setup_ctx(py: Python) -> PyResult<Option<&PyDict>> {
-        let tax = Py::new(
-            py,
-            Taxonomy {
-                t: create_example(),
-            },
-        )?;
-        let ctx = PyDict::new(py);
-        ctx.set_item("tax", tax)?;
-        ctx.set_item("Taxonomy", py.get_type::<Taxonomy>())?;
-        Ok(Some(ctx))
-    }
-
-    #[test]
-    fn test_get() {
-        let gil = Python::acquire_gil();
-        let py = gil.python();
-        let ctx = setup_ctx(py).unwrap();
-
-        let node: TaxonomyNode = py
-            .eval(r#"tax.node("53452")"#, None, ctx)
-            .unwrap()
-            .extract()
-            .unwrap();
-        assert_eq!(node.id, "53452");
-        assert_eq!(node.parent, Some("1046".to_string()));
-        assert_eq!(node.name, "Lamprocystis");
-
-        let node: Option<TaxonomyNode> = py
-            .eval(r#"tax.node("unknown")"#, None, ctx)
-            .unwrap()
-            .extract()
-            .unwrap();
-        assert!(node.is_none());
-    }
-
-    #[test]
-    fn test_find_by_name() {
-        let gil = Python::acquire_gil();
-        let py = gil.python();
-        let ctx = setup_ctx(py).unwrap();
-
-        let node: TaxonomyNode = py
-            .eval(r#"tax.find_by_name("Lamprocystis")"#, None, ctx)
-            .unwrap()
-            .extract()
-            .unwrap();
-        assert_eq!(node.id, "53452");
-        assert_eq!(node.parent, Some("1046".to_string()));
-        assert_eq!(node.name, "Lamprocystis");
-
-        let node: Option<TaxonomyNode> = py
-            .eval(r#"tax.find_by_name("unknown")"#, None, ctx)
-            .unwrap()
-            .extract()
-            .unwrap();
-        assert!(node.is_none());
-    }
-
-    #[test]
-    fn test_parent() {
-        let gil = Python::acquire_gil();
-        let py = gil.python();
-        let ctx = setup_ctx(py).unwrap();
-
-        let parent: TaxonomyNode = py
-            .eval(r#"tax.parent("53452")"#, None, ctx)
-            .unwrap()
-            .extract()
-            .unwrap();
-        assert_eq!(parent.id, "1046");
-
-        let parent: TaxonomyNode = py
-            .eval(r#"tax.parent("53452", at_rank="phylum")"#, None, ctx)
-            .unwrap()
-            .extract()
-            .unwrap();
-        assert_eq!(parent.id, "1224");
-
-        let parent: Option<TaxonomyNode> = py
-            .eval(r#"tax.parent("bad id")"#, None, ctx)
-            .unwrap()
-            .extract()
-            .unwrap();
-        assert!(parent.is_none());
-
-        let parent: Option<TaxonomyNode> = py
-            .eval(r#"tax.parent("bad id", at_rank="species")"#, None, ctx)
-            .unwrap()
-            .extract()
-            .unwrap();
-        assert!(parent.is_none());
-
-        // A bad rank does raise
-        assert!(py
-            .eval(r#"tax.parent("53452", at_rank="bad rank")"#, None, ctx)
-            .is_err());
-    }
-
-    #[test]
-    fn test_children() {
-        let gil = Python::acquire_gil();
-        let py = gil.python();
-        let ctx = setup_ctx(py).unwrap();
-
-        let children: Vec<TaxonomyNode> = py
-            .eval(r#"tax.children("53452")"#, None, ctx)
-            .unwrap()
-            .extract()
-            .unwrap();
-        assert_eq!(
-            children.iter().map(|c| &c.id).collect::<Vec<_>>(),
-            vec!["61598"]
-        );
-    }
-
-    #[test]
-    fn test_lineage() {
-        let gil = Python::acquire_gil();
-        let py = gil.python();
-        let ctx = setup_ctx(py).unwrap();
-
-        let lineage: Vec<TaxonomyNode> = py
-            .eval(r#"tax.lineage("1224")"#, None, ctx)
-            .unwrap()
-            .extract()
-            .unwrap();
-        assert_eq!(
-            lineage.iter().map(|c| &c.id).collect::<Vec<_>>(),
-            vec!["1224", "2", "131567", "1"]
-        );
-    }
-
-    #[test]
-    fn test_parents() {
-        let gil = Python::acquire_gil();
-        let py = gil.python();
-        let ctx = setup_ctx(py).unwrap();
-
-        let lineage: Vec<TaxonomyNode> = py
-            .eval(r#"tax.parents("1224")"#, None, ctx)
-            .unwrap()
-            .extract()
-            .unwrap();
-        assert_eq!(
-            lineage.iter().map(|c| &c.id).collect::<Vec<_>>(),
-            vec!["2", "131567", "1"]
-        );
-    }
-
-    #[test]
-    fn test_lca() {
-        let gil = Python::acquire_gil();
-        let py = gil.python();
-        let ctx = setup_ctx(py).unwrap();
-
-        let lca: TaxonomyNode = py
-            .eval(r#"tax.lca("56812", "765909")"#, None, ctx)
-            .unwrap()
-            .extract()
-            .unwrap();
-        assert_eq!(lca.id, "1236");
-    }
-
-    #[test]
-    fn test_info_methods() -> PyResult<()> {
-        let gil = Python::acquire_gil();
-        let py = gil.python();
-        let ctx = setup_ctx(py)?;
-
-        let root: TaxonomyNode = py.eval(r#"tax.root"#, None, ctx)?.extract()?;
-        assert_eq!(root.id, "1");
-
-        let node: TaxonomyNode = py.eval(r#"tax["1224"]"#, None, ctx)?.extract()?;
-        assert_eq!(node.rank, "phylum");
-        assert_eq!(node.name, "Proteobacteria");
-        Ok(())
-    }
-
-    #[test]
-    fn test_repr() -> PyResult<()> {
-        let gil = Python::acquire_gil();
-        let py = gil.python();
-        let ctx = setup_ctx(py)?;
-
-        let repr: String = py.eval(r#"repr(tax)"#, None, ctx)?.extract()?;
-        println!("{}", repr);
-        assert_eq!(repr, "<Taxonomy (14 nodes)>");
-        Ok(())
-    }
-
-    #[test]
-    fn test_editing() -> PyResult<()> {
-        let gil = Python::acquire_gil();
-        let py = gil.python();
-        let ctx = setup_ctx(py)?;
-
-        let tax: PyRef<Taxonomy> = ctx.unwrap().get_item("tax").unwrap().extract()?;
-        let n_tax_items = TaxTrait::<&str, _>::len(&tax.t);
-        drop(tax);
-
-        assert!(py
-            .eval(r#"tax.add_node("bad id", "new id")"#, None, ctx)
-            .is_err());
-        py.eval(r#"tax.add_node("1236", "91347")"#, None, ctx)?;
-        let tax: PyRef<Taxonomy> = ctx.unwrap().get_item("tax").unwrap().extract()?;
-        assert_eq!(TaxTrait::<&str, _>::len(&tax.t), n_tax_items + 1);
-        drop(tax);
-
-        assert!(py.eval(r#"tax.remove_node("bad id")"#, None, ctx).is_err());
-        py.eval(r#"tax.remove_node("135622")"#, None, ctx)?;
-        let tax: PyRef<Taxonomy> = ctx.unwrap().get_item("tax").unwrap().extract()?;
-        assert_eq!(TaxTrait::<&str, _>::len(&tax.t), n_tax_items);
-        drop(tax);
-
-        assert!(py.eval(r#"del tax["bad id"])"#, None, ctx).is_err());
-        py.run(r#"del tax["1046"]"#, None, ctx)?;
-        let tax: PyRef<Taxonomy> = ctx.unwrap().get_item("tax").unwrap().extract()?;
-        assert_eq!(TaxTrait::<&str, _>::len(&tax.t), n_tax_items - 1);
-        drop(tax);
-
-        // FIXME: make these fail?
-        // assert!(py.eval(r#"tax.prune(keep=["bad id"])"#, None, ctx).is_err());
-        // assert!(py.eval(r#"tax.prune(remove=["bad id"])"#, None, ctx).is_err());
-        py.eval(r#"tax.prune(keep=["22"])"#, None, ctx)?;
-        py.eval(r#"tax.prune(remove=["22"])"#, None, ctx)?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_edit_node() -> PyResult<()> {
-        let gil = Python::acquire_gil();
-        let py = gil.python();
-        let ctx = setup_ctx(py).expect("setup context");
-
-        assert!(py
-            .eval(r#"tax.edit_node("bad id", name="New Name")"#, None, ctx)
-            .is_err());
-        py.eval(r#"tax.edit_node("135622", name="New Name")"#, None, ctx)?;
-        let tax: PyRef<Taxonomy> = ctx.unwrap().get_item("tax").unwrap().extract()?;
-        assert_eq!(tax.t.name("135622").unwrap(), "New Name");
-        drop(tax);
-
-        py.eval(r#"tax.edit_node("135622", rank="genus")"#, None, ctx)?;
-        let tax: PyRef<Taxonomy> = ctx.unwrap().get_item("tax").unwrap().extract()?;
-        assert_eq!(tax.t.rank("135622").unwrap(), TaxRank::Genus);
-        drop(tax);
-
-        assert!(py
-            .eval(r#"tax.edit_node("2", parent_id="135622")"#, None, ctx,)
-            .is_err());
-        assert!(py
-            .eval(r#"tax.edit_node("1", parent_id="131567")"#, None, ctx,)
-            .is_err());
-        assert!(py
-            .eval(r#"tax.edit_node("1", parent_distance=3.5)"#, None, ctx,)
-            .is_err());
-        py.eval(
-            r#"tax.edit_node("135622", parent_id="2", parent_distance=3.5)"#,
-            None,
-            ctx,
-        )?;
-        let tax: PyRef<Taxonomy> = ctx.unwrap().get_item("tax").unwrap().extract()?;
-        assert_eq!(tax.t.parent("135622").unwrap(), Some(("2", 3.5)));
-        drop(tax);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_iteration() -> PyResult<()> {
-        let gil = Python::acquire_gil();
-        let py = gil.python();
-        let ctx = setup_ctx(py)?;
-
-        let n_nodes: u64 = py.eval(r#"len(tax)"#, None, ctx)?.extract()?;
-        assert_eq!(n_nodes, 14);
-
-        let n_nodes: u64 = py.eval(r#"len([t for t in tax])"#, None, ctx)?.extract()?;
-        assert_eq!(n_nodes, 14);
-
-        let contained: bool = py.eval(r#""1224" in tax"#, None, ctx)?.extract()?;
-        assert_eq!(contained, true);
-        let contained: bool = py.eval(r#""nonexistant" in tax"#, None, ctx)?.extract()?;
-        assert_eq!(contained, false);
-        Ok(())
-    }
-
-    #[test]
-    fn test_import() -> PyResult<()> {
-        let gil = Python::acquire_gil();
-        let py = gil.python();
-        let ctx = setup_ctx(py)?;
-
-        let tax: PyRef<Taxonomy> = py
-            .eval(r#"Taxonomy.from_newick("(A,B)C;")"#, None, ctx)?
-            .extract()?;
-        assert_eq!(TaxTrait::<&str, _>::len(&tax.t), 3);
-
-        let tax: PyRef<Taxonomy> = py
-            .eval(
-                r#"Taxonomy.from_json("{\"nodes\": [], \"links\": []}")"#,
-                None,
-                ctx,
-            )?
-            .extract()?;
-        assert_eq!(TaxTrait::<&str, _>::len(&tax.t), 0);
-
-        let tax: PyRef<Taxonomy> = py
-            .eval(r#"Taxonomy.from_json("{\"id\": \"1\"}")"#, None, ctx)?
-            .extract()?;
-        assert_eq!(TaxTrait::<&str, _>::len(&tax.t), 1);
-
-        let tax: PyRef<Taxonomy> = py
-            .eval(
-                r#"Taxonomy.from_json("{\"test\": {\"id\": \"1\"}}", path=["test"])"#,
-                None,
-                ctx,
-            )?
-            .extract()?;
-        assert_eq!(TaxTrait::<&str, _>::len(&tax.t), 1);
-
-        let tax: PyRef<Taxonomy> = py
-            .eval(r#"Taxonomy.from_phyloxml("<phylogeny rooted=\"true\"><clade><id>root</id></clade></phylogeny>")"#, None, ctx)?
-            .extract()?;
-        assert_eq!(TaxTrait::<&str, _>::len(&tax.t), 1);
-
-        let tax: PyRef<Taxonomy> = py
-            .eval(r#"Taxonomy.from_ncbi("tests/data/ncbi_subset_tax.nodes.dmp", "tests/data/ncbi_subset_tax.names.dmp")"#, None, ctx)?
-            .extract()?;
-        assert_eq!(TaxTrait::<&str, _>::len(&tax.t), 9);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_export() -> PyResult<()> {
-        let gil = Python::acquire_gil();
-        let py = gil.python();
-        let ctx = setup_ctx(py)?;
-
-        let json: &[u8] = py.eval(r#"tax.to_json()"#, None, ctx)?.extract()?;
-        assert_eq!(json[0], b'{');
-
-        let nwk: Vec<u8> = py.eval(r#"tax.to_newick()"#, None, ctx)?.extract()?;
-        assert_eq!(nwk[0], b'(');
-        Ok(())
-    }
-
-    #[test]
-    fn test_weights() -> PyResult<()> {
-        let gil = Python::acquire_gil();
-        let py = gil.python();
-        let ctx = setup_ctx(py)?;
-        ctx.map(|ctx| {
-            let w = PyModule::new(py, "weights").unwrap();
-            weights(py, w).unwrap();
-            ctx.set_item("weights", w).unwrap();
-        });
-
-        let rolled: &PyDict = py
-            .eval(
-                r#"weights.rollup_weights(tax, {"1": 1, "2": 4})"#,
-                None,
-                ctx,
-            )?
-            .extract()?;
-        assert_eq!(rolled.get_item("1").unwrap().extract::<f32>()?, 5.);
-        assert_eq!(rolled.get_item("131567").unwrap().extract::<f32>()?, 4.);
-
-        let mwp: (&str, f32) = py
-            .eval(
-                r#"weights.maximum_weighted_path(tax, {"1": 1, "2": 4})"#,
-                None,
-                ctx,
-            )?
-            .extract()?;
-        assert_eq!(mwp.0, "2");
-        assert_eq!(mwp.1, 5.);
-        Ok(())
-    }
 }
